python_app/requirements.txt
paho-mqtt
pymongo
psycopg2-binary
py2neo
